refactor/#86: domain/entity 추가
