+++ app/src/main/java/com/kiero/domain/entity/sse/RawSseEvent.kt
@@ -1,4 +1,4 @@
-package com.kiero.data.sse.model
+package com.kiero.domain.entity.sse
 
 data class RawSseEvent(
     val type: String?,
