+++ app/src/main/java/com/kiero/domain/entity/sse/SseEventType.kt
@@ -1,4 +1,4 @@
-package com.kiero.data.sse.model
+package com.kiero.domain.entity.sse
 
 enum class SseEventType(val value: String) {
     CONNECTED("connected"),
